chore: We split existing services into binaries and crates (#42)
* chore: We split existing services into binaries and crates

We split our services into binaries and crates to make it easy for sharing code.

* fix: docker ci

* chore: all bins as default members
