+++ bin/sidecrush/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "sidecrush-bin"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "sidecrush"
+path = "src/main.rs"
+
+[dependencies]
+clap = { workspace = true }
+sidecrush = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+tokio = { workspace = true }
+cadence = { workspace = true }
