+++ crates/infra/mempool-rebroadcaster/src/lib.rs
@@ -1,1 +1,3 @@
-pub mod rebroadcaster;
+mod rebroadcaster;
+
+pub use rebroadcaster::{Rebroadcaster, RebroadcasterResult, TxpoolDiff};
