+++ crates/infra/sidecrush/src/lib.rs
@@ -1,2 +1,7 @@
-pub mod blockbuilding_healthcheck;
-pub mod metrics;
+mod healthcheck;
+mod metrics;
+
+pub use healthcheck::{
+    BlockProductionHealthChecker, EthClient, HeaderSummary, HealthcheckConfig, Node, alloy_client,
+};
+pub use metrics::HealthcheckMetrics;
